+++ example/bin/mcp_servers.dart
@@ -8,14 +8,14 @@ import 'package:dartantic_ai/dartantic_ai.dart';
 import 'package:logging/logging.dart';
 
 void main() async {
-  Logger.root.level = Level.INFO; // FINE
+  Logger.root.level = Level.FINE; // INFO
   Logger.root.onRecord.listen(
     (record) => print('\n[${record.level.name}]: ${record.message}\n'),
   );
 
-  // await singleMcpServer();
-  // await multipleToolsAndMcpServers();
-  await oneRequestMultiTool(); // simulated calendar MCP server
+  await singleMcpServer();
+  await multipleToolsAndMcpServers();
+  await oneRequestMultiTool();
   exit(0);
 }